# optimize compilation time of argmin/argmax op