<!-- Demo: https://github.com/PaddlePaddle/Paddle/pull/24810 -->
### PR types
<!-- One of [ New features | Bug fixes | Function optimization | Performance optimization | Breaking changes | Others ] -->
Others

### PR changes
<!-- One of [ OPs | APIs | Docs | Others ] -->
OPs

### Describe
<!-- Describe what this PR does -->
cc：
Optimize compilation time of `argmin`: 73s -> 24s
Optimize compilation time of `argmax`: 72s -> 24s
cu：
Optimize compilation time of `argmin`: 85 -> 32s
Optimize compilation time of `argmax`: 84 -> 32s